+++ lib/materials/paper_texture.dart
@@ -5,7 +5,6 @@ import '../theme/wab_theme.dart';
 import '../tokens/material.dart';
 import '../tokens/texture.dart';
 
-/// Paper has two public intensities: ordinary pulp and clouded mottle.
 enum WabPaperTextureKind { paper, mottle }
 
 /// 宣紙紙紋 — deterministic pulp, faint creases, flecks and optional 雲斑.
@@ -44,7 +43,9 @@ class _PaperTexturePainter extends CustomPainter {
 
     final deepTint = isDark ? WAB_PAPER_MOTTLE_DEEP_DARK : WAB_PAPER_MOTTLE_DEEP_LIGHT;
     final paleTint = isDark ? WAB_PAPER_MOTTLE_SHEEN_DARK : WAB_PAPER_MOTTLE_PALE_LIGHT;
-    final mottles = (area10k * (strong ? 1.1 : .42)).round().clamp(3, strong ? 15 : 8);
+    final mottles = (area10k * (strong ? 1.1 : .42))
+        .round()
+        .clamp(3, strong ? 15 : 8);
     for (var i = 0; i < mottles; i++) {
       final rx = size.width * (.09 + rnd.nextDouble() * (strong ? .26 : .17));
       final ry = size.height * (.08 + rnd.nextDouble() * (strong ? .23 : .15));
@@ -58,13 +59,15 @@ class _PaperTexturePainter extends CustomPainter {
           ..color = (i.isEven ? deepTint : paleTint).withOpacity(
             (strong ? .075 : .035) + rnd.nextDouble() * (strong ? .07 : .035),
           )
-          ..maskFilter = MaskFilter.blur(BlurStyle.normal, math.max(8, rx * .48)),
+          ..maskFilter = MaskFilter.blur(BlurStyle.normal, math.max(8.0, rx * .48)),
       );
     }
 
     final crease = isDark ? WAB_TEXTURE_PAPER_CREASE_DARK : WAB_TEXTURE_PAPER_CREASE_LIGHT;
     final highlight = isDark ? WAB_TEXTURE_PAPER_HIGHLIGHT_DARK : WAB_TEXTURE_PAPER_HIGHLIGHT_LIGHT;
-    final creaseCount = (area10k * (strong ? .55 : .28)).round().clamp(2, strong ? 12 : 7);
+    final creaseCount = (area10k * (strong ? .55 : .28))
+        .round()
+        .clamp(2, strong ? 12 : 7);
     for (var i = 0; i < creaseCount; i++) {
       final x0 = rnd.nextDouble() * size.width;
       final y0 = rnd.nextDouble() * size.height;
@@ -72,8 +75,12 @@ class _PaperTexturePainter extends CustomPainter {
       var x = x0;
       var y = y0;
       for (var j = 0; j < 3; j++) {
-        x = (x + (rnd.nextDouble() - .5) * size.width * .18).clamp(0, size.width);
-        y = (y + (rnd.nextDouble() - .5) * size.height * .20).clamp(0, size.height);
+        x = (x + (rnd.nextDouble() - .5) * size.width * .18)
+            .clamp(0.0, size.width)
+            .toDouble();
+        y = (y + (rnd.nextDouble() - .5) * size.height * .20)
+            .clamp(0.0, size.height)
+            .toDouble();
         path.lineTo(x, y);
       }
       canvas.drawPath(
@@ -95,7 +102,8 @@ class _PaperTexturePainter extends CustomPainter {
       canvas.restore();
     }
 
-    final dots = (area10k * (isDark ? WAB_PAPER_DOT_DENSITY_DARK : WAB_PAPER_DOT_DENSITY_LIGHT))
+    final dots = (area10k *
+            (isDark ? WAB_PAPER_DOT_DENSITY_DARK : WAB_PAPER_DOT_DENSITY_LIGHT))
         .round();
     final dotPaint = Paint()
       ..color = (isDark ? WAB_PAPER_TINT_DARK : WAB_PAPER_DOT_TINT_LIGHT)
